refactor: split playlist.rs into auto_sync/dynamic/import_export submodules (#577 item 18)
Splits three self-contained domains out of playlist.rs's 2857-line
monolith, each as a second `impl PlaylistManager` block in its own
file (same pattern as #577 item 17's collection.rs split):

- playlist/auto_sync.rs: system-managed genre/decade/BPM auto-playlist
  sync (sync_all/genre/decade/bpm_auto_playlists)
- playlist/dynamic.rs: dynamic/Smart Playlist spec dispatch, populate,
  and library-change reconciliation (songs_for_spec,
  populate_dynamic_playlist, reconcile_dynamic_playlists, etc.)
- playlist/import_export.rs: .m3u/.pls import and export

playlist.rs itself is now 1313 lines, holding playlist CRUD, item
mutation, and the undo/redo stack — kept together since undo/redo is
tightly coupled to the item-mutation methods (shared touch_updated/
reorder_item_internal/renumber_positions helpers).

No behavior change — pure move. 218/219 tests pass throughout,
clippy/fmt clean (12 pre-existing unrelated findings in player.rs/
tags.rs, same as prior PRs in this series).
